Add LicenceDocumentModel from crm_v2.documents (#618)
https://eaflood.atlassian.net/browse/WATER-4292

> This is part of a series of changes to add the ability to identify the licence holder for a licence

This change adds the migration, model, test helper and unit tests for a new `LicenceDocumentModel`.

TBH, it is a pointless model! This seems to be a bridge between the `water.licences` table we now have and the `crm.document` that was created back when the idea was to have a generic 'repository' for all types of documents, not just water abstraction licences. We know the schema `crm_v2` is the result of an incomplete attempt to move away from that idea. But for whatever reason, rather than link the `water.licences` records directly to contact and company information instead the previous team stuck with the concept of a 'document'

So, we need to add this model to then be able to identify the relevant `crm_v2.document_roles` records we need, to find the current licence holder! 🥳😩