Add LicenceDocument relationship to LicenceModel
https://eaflood.atlassian.net/browse/WATER-4292

> This is part of a series of changes to add the ability to identify the licence holder for a licence

This change adds the migration, model, test helper and unit tests for a new `LicenceDocumentRoleModel`.

We added [LicenceDocumentModel](https://github.com/DEFRA/water-abstraction-system/pull/618). Though pointless, it's the only thing that points us to who the licence holder is. We then added [LicenceRoleModel](https://github.com/DEFRA/water-abstraction-system/pull/629) and [LicenceDocumentRoleModel](https://github.com/DEFRA/water-abstraction-system/pull/630) which in combination with `LicenceDocumentModel` will give us the licence holder.

But we intend to work with `LicenceModel` instances. `LicenceDocumentModel` is just a really, _really_ cut down copy of the licence record which the previous team added for 'reasons'. So, this last change before we can get on and identify the licence holder is to setup the relationship between `LicenceModel` and `LicenceDocumentModel`.