Fix missing bill in view bill run screen
https://eaflood.atlassian.net/browse/WATER-4223

We are replacing the legacy bill run view with one we have developed in this project. During testing it was found that there are examples of bill runs where a bill is missing from the view. When we checked the legacy version the bill was present so this is an issue just in our version of the page.

This change fixes the feature to ensure all bills are displayed in the view as expected.