Left join to BillingAccountAddresses
Having dug into the examples provided the root cause is that the missing bills in the view are linked to billing accounts that _do not_ have billing account address records.

Till now we have been working with the assumption that all billing accounts must have at least 1 billing account record. It turns out that is not the case.

We've tweaked our unit test setup for `FetchBillRunService` to ensure this scenario is covered and then amended the service to use a left join.

But this opens a can of worms with the rest of the bill run pages we'll need to look into.