Identify licence holder
We've had to make quite a jump in complexity but there is not way round it thanks to how the previous team chose to store licence details.

We make use of the new Licence document, role and document-role models we recently created to get to the company and/or contact record that is identified as the licence holder.

We then have to apply a bit of logic to determine which 'party' to use; the company or the contact. Contact takes precedence. We've already had to deal with the madness of displaying a contact name, so can take advantage of the existing model instance method `$name()` to sort that out for us.