Update to use DB Id rather than pre-generated one
We can avoid bringing in an extra module and let the DB generate the ID by using [returning()](https://vincit.github.io/objection.js/api/query-builder/find-methods.html#returning).

We initially avoided it because it gave the impression that it ignores what we ask the DB to return and instead returns everything. But [this issue](https://github.com/Vincit/objection.js/issues/1773) and our own testing confirmed Objection.js is just combining what we passed in with what we asked the DB to return, which gives the impression the DB is sending back more than we wanted.