Housekeeping of return requirement routes
https://eaflood.atlassian.net/browse/WATER-4261

This is just a bit of housekeeping of the return requirements setup routes before things get too baked in.

Key thing is the ordering of things in the files; we have a convention of doing things alphabetically so they could do with being tidied up.

We've also spotted that some routes have a `returns-` prefix and others don't. So, this will make them consistent.

Finally, this change proposes renaming the controller handlers that deal with the POST requests to be prefixed with `submit` rather than `save`. Our reasoning is they are handling the form _submission_, and should there be a validation issue no save will take place.