Update /approved to expect licence ID
We expect to persist the return requirement and delete the session when the 'Approved' button is clicked in the check your answers page.

So, we won't have a session record to fetch and use. But we will need to know which licence we approved the return requirement for so we can set the link needed in the page.

This updates the route to include that and adds the link to the view so we have the context of why we need it.