Housekeeping of return requirement routes (#647)
https://eaflood.atlassian.net/browse/WATER-4261

This is some housekeeping of the return requirements setup routes before things get too baked in.

Primarily it started as sorting the order of the handlers, routes and tests to be alphabetical. However, once we dug in we spotted several other things that needed sorting.

- make the endpoint paths consistent
- make handler names consistent with path names
- add missing `POST` endpoints
- add missing `POST` handlers
- redirect instead of return in all `POST` handlers
- rename `POST` handlers from `save` to `submit` (reason being they are handling the form _submission_, and if there is a validation issue no save will take place)
- ensure licence ID gets through to the `/approve` page
- make route descriptions consistent
- refactor tests to reuse `options`

So, it's a bigger change than we'd normally do. But a number are pretty intrinsically linked so we're bundling them into this single housekeeping change.