Support both return requirements set up journeys
https://eaflood.atlassian.net/browse/WATER-4261

The return requirement setup journey is backed by a `SessionModel` instance that is used to store the values a users enters as well as helpful info like the licence ID and reference. But at the moment we are only creating it for the 'no returns required' journey. We really need the session to be initialised for both journeys in order to develop and test all the pages needed.

So, this change adds a new endpoint to `/licences` that will do the same thing as `LicencesController.noReturnsRequired()`. Only both handlers will now pass a value that denotes the journey type to the `InitiateReturnRequirementSessionService` so it can store it in the session for future reference.