Fix return requirements start-date back link
https://eaflood.atlassian.net/browse/WATER-4261

We're about to begin work on [adding the controls and validation](https://github.com/DEFRA/water-abstraction-system/pull/646) for `/start-date` page in the returns requirements set up journey.

But we've also added [a test-only section](https://github.com/DEFRA/water-abstraction-ui/pull/2490) to the legacy UI that allows us to start both set up journeys and stop having to fudge the URL's manually.

But making it easy to start the journey has highlighted when we try to go back we're going to the wrong place. We want our QA team to be able to crack on with testing. So, though we know this page is due to be implemented properly it will cause us less headaches if we can quickly sort this issue in the meantime.