Fix missing titles in returns reqs. set up pages
https://eaflood.atlassian.net/browse/WATER-4261

Whilst working on some [housekeeping](https://github.com/DEFRA/water-abstraction-system/pull/647) and [tweaks](https://github.com/DEFRA/water-abstraction-system/pull/648) we spotted that the titles for all the new returns requirements set up pages were missing.

A quick check identified that all the pages are using our standard `layout.njk` (😁) but it expects a `pageTitle` value to be passed in the context and none of the pages are doing this (😱).

This changes fixes the issue and adds the missing titles.