Add page title for no returns required page
This is the first page that has had controls and validation added to it. As such it has adopted the pattern we expect the other placeholder pages to follow when they get updated.

We were initially going to add the page title to `NoReturnsRequiredPresenter`. But then we realised the concept of presenters is to take some existing data and transform it ready for 'presenting' elsewhere. Returning a value that hasn't been generated from the data passed in breaks that concept.

So, we're also tweaking things a little here. We add the title directly to what the service returns. We also move `activeNavBar:` out of the presenter because it too is a break from what how we intend to use presenters.