Boy scout tidy up
First, we rename `_radioButtons()` to be `_radioItems()` as that is what it is generating and the property it is setting.

We then spotted that `session` wasn't being used in `NoReturnsRequiredPresenter._error()` so removed the argument. For `NoReturnsRequiredPresenter.__radioItems()` we do intend to use it in future because it will tell us which item has previously been selected. So, we don't remove it, but do indicate we are not using it by applying the 'underscore' convention.