Remove no returns check your answers page
We were about to sort out the title for the 'Check your answers page'. But then we remembered we have 2 of them so we'd need to do it in 2 places.

We have no intention of maintaining 2 versions of the same page. So, rather than duplicate something again we use this opportunity to get shot of the no returns check your answers page, route and controller handlers.