Boy scout - Remove set title from views
When we updated the title for 'Check your answers' we spotted that the title was actually be set to a variable in the template and then referred to in the H1. We're not sure what benefit that adds over just putting the title directly into the H1.

So, this helps us get rid of some lines of code and complexity from our views.