Fix check your answers for tests
We broke the test because there is no session object with a licence backing the tests at the moment. In the future fetching the session and formatting the data for check your answers will be done by a service that we can then stub in the controller test.

But as we're not ready for that yet (and we still haven't figured out how to stub Objection model queries 😢) we 'fudge' it and use the javascript safe navigation operator.