Add LicenceVersionModel from water.licence_versions
https://eaflood.atlassian.net/browse/WATER-4261

> This change supports the return requirements setup work

Tickets WATER-4263 and WATER-4266 covers adding controls and validation to the start date page in the set up journeys (returns required and not required).

Users will be offered the chose to choose an existing start date or enter one manually. The existing start date needs to come from the current licence version record.

So, as a first step to enabling that we need to add the `LicenceVersionModel` to our project and link it to `LicenceModel`.