Determine licence start date 4 returns reqs setup
https://eaflood.atlassian.net/browse/WATER-4261

> This change supports the return requirements setup work

Tickets WATER-4263 and WATER-4266 cover adding controls and validation to the start date page in the set-up journeys (returns required and not required).

Users will be offered the choice to choose an existing start date or enter one manually. The existing start date needs to come from the current licence version record.

So, this change adds functionality to the `InitiateReturnRequirementSessionService` to add the start date to the session for use later in the page.