Refactor Persist Allocated Licences To Results Service
https://eaflood.atlassian.net/browse/WATER-4319

As part of the work we have been doing, we've been hacking away at a branch in a mob to build the two-part-tariff service. We have started refactoring this hacky branch and pulling out the various services into their own PR's. DEFRA/water-abstraction-team#105. Since then we have found a much cleaner, more performant way to right the overall service that we have now added into another separate hacky branch. This PR is updating the PersistAllocatedLicencesToResultsService that has already been pulled out and refactoring it to match our new hacky branch.