Refactor Prepare Licence For Allocation Service (#656)
https://eaflood.atlassian.net/browse/WATER-4319

This pull request refactors the 'PrepareLicenceForAllocationService' as part of our ongoing effort to enhance the two-part-tariff service. Initially we began by creating a 'hack' branch to build and iterate on the two-part-tariff functionality. In the process of working on these branches, we identified a cleaner and more performant approach to implement the overall service. We realised that the PrepareLicenceForAllocationService could be split into two smaller services, making it easier in the code to follow the two-part-tariff process and making it easier for testing.