Refactor Persist Allocated Licences To Results Service (#658)
https://eaflood.atlassian.net/browse/WATER-4319

This pull request refactors the 'PersistAllocatedLicencesToResultsService' as part of our ongoing effort to enhance the two-part-tariff service. Initially, we began by creating a 'hack' branch to build and iterate on the two-part-tariff functionality. In the process of working on these branches, we identified a cleaner and more performant approach to implement the overall service.  We realised that the PersistAllocatedLicencesToResultsService would be much cleaner and easier to follow if it only persisted one licence at a time, rather than iterating through every licence. This also makes it easier for testing.