Add LicenceDocumentHeaderModel (#640)
https://eaflood.atlassian.net/browse/WATER-4322

> This is part of a series of changes to allow us to replace the legacy view licence page

When viewing a licence using the [legacy water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) the summary tab features several links to other pages, some of which give expanded details for the item shown.

Those links rely on having the `crm.document_header` ID for the licence currently being viewed. Our work on WATER-4292 (identifying the licence holder) has shown that much of the data in the table has already been imported elsewhere.

But without the ID those links and pages won't work and they are beyond the scope of what we are doing with the view licence page.

So, to allow us to quickly get the related `crm.document_header` (that we have chosen to call `LicenceDocumentHeader`) for a licence this change adds the model to the project.