Allocate returns to licence service (#628)
https://eaflood.atlassian.net/browse/WATER-4188

This pull request refactors the 'AllocateReturnsToLicenceService' as part of our ongoing effort to enhance the two-part tariff service. Initially, we began by creating a 'hack' branch to build and iterate on the two-part-tariff functionality. In the process of working on these branches, we identified a cleaner and more performant approach to implement the overall service. We realised that the AllocateReturnsToLicenceService would be much cleaner and easier to follow if it was split into two services the AllocateReturnsToLicenceService and the MatchReturnsToChargeElementService.
We have also refactored the allocation service to only allocate one charge element at a time from a licence, rather than iterating through every licence and every charge element. This also makes it easier for testing.