Fix and improve acceptance testing attributes (#665)
https://eaflood.atlassian.net/browse/WATER-4336

In WATER-4070 we logged that the Billing & Data team struggled because they could not access the larger bill runs. The service would appear to be busy for a while and then eventually error.

We figured out it was simply that the service could not return all the data needed using the current design, which listed out all the transactions in a bill run. We also knew the data could be collected in a more performant way.

So, we ended up rebuilding the whole journey of viewing a bill run, the bills in it and the transaction details.

This also allowed us to add attributes to elements likely referenced in our acceptance tests. A [recognised best practice in selecting elements in tests](https://docs.cypress.io/guides/references/best-practices#Selecting-Elements) is to avoid brittle, complex selectors and to use `data-*` attributes instead.

> For reference we have opted for `data-test` as our convention

The legacy pages had none of this, and often didn't even bother with IDs. This means a common battle with the acceptance tests we inherited is finding suitable selectors for the elements we needed.

We now need to [update those tests we broke](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/78) because of our new pages. Doing this has allowed us to spot a few places where we've made mistakes and others where `data-test` would help. This change is us making those tweaks to the billing templates.