Improve tear down speed (#671)
https://eaflood.atlassian.net/browse/WATER-4336

The tear-down feature supports our [water-abstraction-acceptance-tests](https://github.com/DEFRA/water-abstraction-acceptance-tests). At the start of each test, it will make an HTTP call to `POST /data/tear-down`, resulting in `TearDownService.go()` being called.

The service and those it calls are responsible for clearing the various schemas we use of test data. Some tables have a `is_test` field that immediately identifies the record as being created for a test.

But each test will create data that won't be flagged this way. So, the tear-down services also remove records we can identify as being test data.

- it might be linked to a `is_test` record
- it has a known value, for example, `name`, that is specific to the acceptance tests

Ideally, we wouldn't need to remove anything between the tests. We could then avoid this step and the delay it causes. But until we re-architect fully the acceptance tests we've inherited we have to work with what we've got.

When we took on WRLS [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) had a tear-down endpoint. But it would take more than a minute to complete. This caused the tests to take a long time to finish and often caused them to fail.

We built a more performant version. It would typically get the job done in 15 to 20 seconds. But sometimes this would creep to 40 seconds which again would cause test runs to fail.

So, we've gone back and tried a series of alternate strategies for clearing the data. Luckily, we've been able to identify a faster solution. This change updates tear-down to use the new strategy. We've also included a README covering what we tried and the results that led us to this decision (repeated below).

> The average times are in milliseconds. Just divide by 1000 for secs, for example, 1770 / 1000 = 1.7 secs

| Schema  | No change | Just triggers | Single Query | Promise all | Best of | All in one |
|---------|-----------|---------------|--------------|-------------|---------|------------|
| CRM     | 1770      | 2106          | 1185         | **964**     | 1444    | N/A        |
| IDM     | 58        | 73            | 46           | **852**     | 1199    | N/A        |
| PERMIT  | 165       | 266           | 156          | **893**     | 1282    | N/A        |
| RETURNS | 7176      | 6121          | 3517         | **4474**    | 7528    | N/A        |
| WATER   | 42268     | 4413          | 5266         | **5106**    | 5289    | N/A        |
| -       | -         | -             | -            | -           | -       |            |
| Total   | 42269     | 6556          | 5266         | **5106**    | 7529    | 7457       |